% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\alias{dispersion.data.frame}
\title{Calculate Cross-Sectional Dispersion Measures}
\usage{
dispersion(df, method)

\method{dispersion}{data.frame}(df, method = IQR)
}
\arguments{
\item{df}{a \code{data.frame} holding the SPF data.}

\item{method}{a function specifying the dispersion measure, e.g. \code{IQR}, \code{sd}.}
}
\value{
a \code{tsibble}
}
\description{
Calculate Cross-Sectional Dispersion Measures
}
\details{
Computes dispersion measures for the SPF datasets.
}
\examples{
\donttest{
tmp <- get_individual(survey = "CPI")
dsp <- dispersion(tmp, IQR)
}
}
\seealso{
\link{get_individual}
}
