% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{get_additional_cpie10}
\alias{get_additional_cpie10}
\title{Additional CPIE10}
\usage{
get_additional_cpie10(file, row.names = FALSE, ...)
}
\arguments{
\item{file}{if specified, \code{write.csv} writes the dataset to the disk with name \code{file}.}

\item{row.names}{indicates whether the \code{write.csv} should write \code{row.names}.}

\item{...}{further arguments passed to \code{write.csv}.}
}
\description{
Download the additional CPIE10 dataset.
}
