% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anxious_index.R
\name{get_anxious_index}
\alias{get_anxious_index}
\title{Anxious Index}
\usage{
get_anxious_index(file, row.names = FALSE, ...)
}
\arguments{
\item{file}{if specified, \code{write.csv} writes the dataset to the disk with name \code{file}.}

\item{row.names}{indicates whether the \code{write.csv} should write \code{row.names}.}

\item{...}{further arguments passed to \code{write.csv}.}
}
\description{
Download the Anxious Index dataset.
}
