% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dispersion.R, R/get_individual.R,
%   R/get_mean.R, R/get_median.R
\name{get_dispersion}
\alias{get_dispersion}
\alias{get_individual}
\alias{get_mean}
\alias{get_median}
\title{Download SPF Data}
\usage{
get_dispersion(survey, file, row.names = FALSE, ...)

get_individual(survey, file, row.names = FALSE, ...)

get_mean(survey, type = "level", file, row.names = FALSE, ...)

get_median(survey, type = "level", file, row.names = FALSE, ...)
}
\arguments{
\item{survey}{The survey you want to download, e.g. "NGDP", "CPI", prgdp".}

\item{file}{path, the data gets written to, e.g. "file.csv"}

\item{row.names, }{passed to \code{write.csv}.}

\item{...}{further arguments passed to \code{write.csv}.}

\item{type}{one of level: "level", change: "pct", or "ave".}
}
\value{
Returns a \code{data.frame}.
}
\description{
Download a dataaset from the Survey of Professional Forecasters (SPF) provided by the Federal Reserve Bank of Philadelphia.
}
\details{
Download data from the \href{https://www.philadelphiafed.org/research-and-data/real-time-center/survey-of-professional-forecasters}{Survey of Professional Forecasters provided by the Federal Reserve Bank of Philadelphia}.
}
\section{Functions}{
\itemize{
\item \code{get_dispersion}: Download disersion SPF data

\item \code{get_mean}: Download mean SPF data

\item \code{get_median}: Download median SPF data
}}

\examples{
\dontrun{
get_individual("EMP")
get_mean("RGDP", type = "growth")
get_mean("EMP", type = "level")
get_mean("EMP", type = "pct")
get_mean("EMP", type = "ave")
get_median("EMP", type = "level")
get_median("EMP", type = "pct")
get_median("EMP", type = "ave")
get_dispersion("EMP")
}
}
\seealso{
\code{\link[utils:write.table]{utils::write.csv()}}
}
