% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inflation.R
\name{get_inflation}
\alias{get_inflation}
\title{Inflation}
\usage{
get_inflation(file, row.names = FALSE, ...)
}
\arguments{
\item{file}{if specified, \code{write.csv} writes the dataset to the disk with name \code{file}.}

\item{row.names}{indicates whether the \code{write.csv} should write \code{row.names}.}

\item{...}{further arguments passed to \code{write.csv}.}
}
\description{
Download the Inflation dataset.
}
