% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_microdata.R
\name{get_microdata}
\alias{get_microdata}
\title{Download SPFmicrodata as Excel file}
\usage{
get_microdata(file = "SPFmicrodata.xlsx")
}
\arguments{
\item{file}{file name}
}
\description{
Download the SPFmicrodata.csv dataset.
}
