% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob.R
\name{get_prob}
\alias{get_prob}
\title{Download probability data as Excel file}
\usage{
get_prob(file = "prob.xlsx")
}
\arguments{
\item{file}{file name}
}
\description{
Download the the probability dataset.
}
