% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_group.R
\name{summarize_group}
\alias{summarize_group}
\alias{summarize_group.data.frame}
\title{Calculate Dispersion Measures}
\usage{
summarize_group(df, f, ...)

\method{summarize_group}{data.frame}(df, f = IQR, ...)
}
\arguments{
\item{df}{a \code{data.frame} holding the SPF, or other data.}

\item{f}{a function specifying the dispersion measure, e.g. \code{IQR}, \code{sd}. Defaults to \code{IQR}.}

\item{...}{additional arguments passed to \code{f}.}
}
\value{
a \code{tsibble}
}
\description{
Calculate Dispersion Measures
}
\details{
Computes dispersion measures for the SPF datasets.
}
\examples{
\donttest{
tmp <- get_individual(survey = "CPI")
dsp <- summarize_group(tmp, IQR)
}
}
\seealso{
\link{get_individual}
}
